Refactor : GetLatest() and UpdateLatest() now use repo_wrappers.
- Note: If you're thinking the placement of the methods seems a bit awkward, I did it like this to avoid cyclic imports.
